restructured the client classes, agent verbs
