+++ python/samples/concepts/realtime/simple_realtime_chat_websocket.py
@@ -7,8 +7,8 @@ from samples.concepts.realtime.utils import AudioPlayerWebsocket, AudioRecorderW
 from semantic_kernel.connectors.ai.open_ai import (
     AzureRealtimeExecutionSettings,
     AzureRealtimeWebsocket,
-    ListenEvents,
 )
+from semantic_kernel.connectors.ai.open_ai.services._open_ai_realtime import ListenEvents
 from semantic_kernel.contents import RealtimeAudioEvent, RealtimeTextEvent
 
 logging.basicConfig(level=logging.WARNING)
@@ -25,7 +25,7 @@ It requires installing:
 - pyaudio
 - sounddevice
 - pydub
-e.g. pip install pyaudio sounddevice pydub semantic_kernel[realtime]
+e.g. pip install pyaudio sounddevice pydub semantic-kernel[realtime]
 
 For more details of the exact setup, see the README.md in the realtime folder.
 """
@@ -43,10 +43,6 @@ async def main() -> None:
     # create the realtime client and optionally add the audio output function, this is optional
     # you can define the protocol to use, either "websocket" or "webrtc"
     # they will behave the same way, even though the underlying protocol is quite different
-    realtime_client = AzureRealtimeWebsocket()
-    audio_player = AudioPlayerWebsocket()
-    audio_recorder = AudioRecorderWebsocket(realtime_client=realtime_client)
-    # Create the settings for the session
     settings = AzureRealtimeExecutionSettings(
         instructions="""
     You are a chat bot. Your name is Mosscap and
@@ -61,8 +57,12 @@ async def main() -> None:
         # for more details.
         voice="shimmer",
     )
+    realtime_client = AzureRealtimeWebsocket(settings=settings)
+    audio_player = AudioPlayerWebsocket()
+    audio_recorder = AudioRecorderWebsocket(realtime_client=realtime_client)
+    # Create the settings for the session
     # the context manager calls the create_session method on the client and starts listening to the audio stream
-    async with audio_player, audio_recorder, realtime_client(settings=settings, create_response=True):
+    async with audio_player, audio_recorder, realtime_client:
         async for event in realtime_client.receive():
             match event:
                 # this can be used as an alternative to the callback function used in other samples,
@@ -82,9 +82,8 @@ async def main() -> None:
 
 if __name__ == "__main__":
     print(
-        "Instructions: The model will start speaking immediately,"
-        "this can be turned off by removing `create_response=True` above."
-        "The model will detect when you stop and automatically generate a response. "
+        "Instructions: Start speaking. "
+        "The model will detect when you stop and automatically start responding. "
         "Press ctrl + c to stop the program."
     )
     asyncio.run(main())
